refactor: 提取 TouchInputHandler.java 处理触控/鼠标/触控笔事件
从 Game.java 提取 ~1045 行触控处理代码到独立的 TouchInputHandler 类:
- handleMotionEvent: 鼠标/触屏/触控笔的全部 MotionEvent 处理
- updateMousePosition: 鼠标位置跟踪与外接显示器坐标映射
- 触控笔/原生触控手势: trySendPenEvent, trySendTouchEvent
- 多指手势检测: 双指右键, 多指轻触
- 触控上下文管理: setTouchMode, setEnhancedTouch, initTouchContexts
- 华为鼠标中键/滚轮模拟逻辑

Game.java: 4913 → 3677 行 (减少 1236 行)
